+++ CHANGELOG.md
@@ -71,6 +71,7 @@
 * Guess ImGui input window size so that it is not truncated during encode
 * Tab key now work as hotkey when input editor has focus
 * Fix pointer scan for 32-bit games
+* Fix audio mixing with non-integer samples per frame (#600)
 
 ## [1.4.5] - 2023-10-22
 ### Added